requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
aiosqlite
pydantic-settings
jinja2
python-multipart
coinbase-advanced-py
tablib[xlsx,ods]
openpyxl
odfpy
cryptography
python-jose[cryptography]
passlib[bcrypt]
pyotp
qrcode[pil]
slowapi
psutil
httpx
websockets
